[OPIK-4222] [BE] Revision 20: Extract EvaluatorType JDBI classes to infrastructure/db
