+++ apps/opik-backend/src/main/java/com/comet/opik/domain/DatasetEvaluatorDAO.java
@@ -1,15 +1,11 @@
 package com.comet.opik.domain;
 
 import com.comet.opik.api.DatasetEvaluator;
-import com.comet.opik.api.EvaluatorType;
+import com.comet.opik.infrastructure.db.EvaluatorTypeArgumentFactory;
+import com.comet.opik.infrastructure.db.EvaluatorTypeColumnMapper;
 import com.comet.opik.infrastructure.db.JsonNodeArgumentFactory;
 import com.comet.opik.infrastructure.db.JsonNodeColumnMapper;
 import com.comet.opik.infrastructure.db.UUIDArgumentFactory;
-import org.jdbi.v3.core.argument.AbstractArgumentFactory;
-import org.jdbi.v3.core.argument.Argument;
-import org.jdbi.v3.core.config.ConfigRegistry;
-import org.jdbi.v3.core.mapper.ColumnMapper;
-import org.jdbi.v3.core.statement.StatementContext;
 import org.jdbi.v3.sqlobject.config.RegisterArgumentFactory;
 import org.jdbi.v3.sqlobject.config.RegisterColumnMapper;
 import org.jdbi.v3.sqlobject.config.RegisterConstructorMapper;
@@ -20,18 +16,15 @@ import org.jdbi.v3.sqlobject.statement.SqlBatch;
 import org.jdbi.v3.sqlobject.statement.SqlQuery;
 import org.jdbi.v3.sqlobject.statement.SqlUpdate;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Types;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 
 @RegisterArgumentFactory(UUIDArgumentFactory.class)
 @RegisterArgumentFactory(JsonNodeArgumentFactory.class)
-@RegisterArgumentFactory(DatasetEvaluatorDAO.EvaluatorTypeArgumentFactory.class)
+@RegisterArgumentFactory(EvaluatorTypeArgumentFactory.class)
 @RegisterColumnMapper(JsonNodeColumnMapper.class)
-@RegisterColumnMapper(DatasetEvaluatorDAO.EvaluatorTypeColumnMapper.class)
+@RegisterColumnMapper(EvaluatorTypeColumnMapper.class)
 @RegisterConstructorMapper(DatasetEvaluator.class)
 public interface DatasetEvaluatorDAO {
 
@@ -59,35 +52,4 @@ public interface DatasetEvaluatorDAO {
 
     @SqlUpdate("DELETE FROM dataset_evaluators WHERE workspace_id = :workspaceId AND dataset_id = :datasetId")
     int deleteByDatasetId(@Bind("workspaceId") String workspaceId, @Bind("datasetId") UUID datasetId);
-
-    class EvaluatorTypeArgumentFactory extends AbstractArgumentFactory<EvaluatorType> {
-        public EvaluatorTypeArgumentFactory() {
-            super(Types.VARCHAR);
-        }
-
-        @Override
-        protected Argument build(EvaluatorType value, ConfigRegistry config) {
-            return (position, statement, ctx) -> {
-                if (value == null) {
-                    statement.setNull(position, Types.VARCHAR);
-                } else {
-                    statement.setString(position, value.getValue());
-                }
-            };
-        }
-    }
-
-    class EvaluatorTypeColumnMapper implements ColumnMapper<EvaluatorType> {
-        @Override
-        public EvaluatorType map(ResultSet r, int columnNumber, StatementContext ctx) throws SQLException {
-            String value = r.getString(columnNumber);
-            return value == null ? null : EvaluatorType.fromString(value);
-        }
-
-        @Override
-        public EvaluatorType map(ResultSet r, String columnLabel, StatementContext ctx) throws SQLException {
-            String value = r.getString(columnLabel);
-            return value == null ? null : EvaluatorType.fromString(value);
-        }
-    }
 }
